doc: document frequency dividing
* Documented frequency dividing.
* Moved frequency dividing to separate file.

Activity: #19
